Organize typeDefs
