+++ graphql/typeDefs/base.js
@@ -1,0 +1,15 @@
+const { gql } = require('apollo-server')
+
+module.exports = gql`
+  type Query {
+    _: String
+  }
+
+  type Mutation {
+    _: String
+  }
+
+  type Subscription {
+    _: String
+  }
+`
