+++ graphql/typeDefs/index.js
@@ -1,0 +1,5 @@
+const base = require('./base')
+const post = require('./post')
+const user = require('./user')
+
+module.exports = [base, post, user]
